pedestrians-cluster/requirements.txt
tqdm
numpy>=1.18.1
scipy
sklearn
infomap
faiss-gpu
torch>=1.2
torchvision